Move the applyThousandSeparator method to utils (#4576)
* Remove applyThousandSeparator from index file

* Create the utils.js file

* Move applyThousandSeparator to the utils file

* Use applyThousandSeparator in the index file

* Update changelog

* Move applyThousandSeparator to client/utils/index.js file

* Use applyThousandSeparator in transactions count

* Use applyThousandSeparator in disputes count

* Use applyThousandSeparator in deposits count

* Update changelog

* Remove previous changelog
