+++ client/transactions/list/index.tsx
@@ -4,7 +4,7 @@
  * External dependencies
  */
 import React, { Fragment, useState } from 'react';
-import { uniq, partial } from 'lodash';
+import { uniq } from 'lodash';
 import { useDispatch } from '@wordpress/data';
 import { useMemo } from '@wordpress/element';
 import { dateI18n } from '@wordpress/date';
@@ -27,7 +27,6 @@ import {
 	generateCSVFileName,
 } from '@woocommerce/csv-export';
 import apiFetch from '@wordpress/api-fetch';
-import { numberFormat } from '@woocommerce/number';
 
 /**
  * Internal dependencies
@@ -53,6 +52,8 @@ import wcpayTracks from 'tracks';
 import DownloadButton from 'components/download-button';
 import { getTransactionsCSV } from '../../data/transactions/resolvers';
 import p24BankList from '../../payment-details/payment-method/p24/bank-list';
+import { applyThousandSeparator } from '../../utils/index.js';
+
 interface TransactionsListProps {
 	depositId?: string;
 }
@@ -77,28 +78,6 @@ interface Column extends TableCardColumn {
 	visible?: boolean;
 	cellClassName?: string;
 }
-
-const applyThousandSeparator = ( trxCount: number ) => {
-	const siteLang = document.documentElement.lang;
-	const siteNumberOptions = {
-		thousandSeparator: ',',
-	};
-
-	if ( [ 'fr', 'pl' ].some( ( lang ) => siteLang.startsWith( lang ) ) ) {
-		siteNumberOptions.thousandSeparator = ' ';
-	} else if ( 'de-CH' === siteLang ) {
-		siteNumberOptions.thousandSeparator = "'";
-	} else if (
-		[ 'de', 'nl', 'it', 'es', 'pt' ].some( ( lang ) =>
-			siteLang.startsWith( lang )
-		)
-	) {
-		siteNumberOptions.thousandSeparator = '.';
-	}
-
-	const formattedNumber = partial( numberFormat, siteNumberOptions );
-	return formattedNumber( trxCount );
-};
 
 const getPaymentSourceDetails = ( txn: Transaction ) => {
 	if ( ! txn.source_identifier ) {
